No need for koharness to require kohana/core
koharness itself doesn't need it's own version of kohana/core - this would normally be 
a dependency of the module being tested. Removed it from the composer.json
to avoid circular dependency issues as experienced by @lenton on 
kohana/core#548

There are a couple of classes provided by koharness that will fail if kohana/core
is not present, but that's such an edge case it's ok to trust developers to figure 
that out.